+++ strider/__main__.py
@@ -1,91 +1,30 @@
+# todo split this file it is gigantic
+
 import argparse
 import textwrap
-import importlib
+import importlib.util
 import warnings
 import sys
+import json
 import itertools as it
-from collections import ChainMap
-from collections.abc import Mapping
+import os.path as path
 
 import cv2
 import strider
+from strider import TrackPack
+from strider.rectangle import Rectangle
 from strider.__util__ import *
 
 if strider.__dev__:
     print('you are using a development version of strider, install and use a release version unless'
           ' you know what you are doing', file=sys.stderr)
 
-
-class Calibrate(argparse.Action):
-    # calibrate action
-    def __init__(self, *args, **kwargs):
-        kwargs.setdefault('default', argparse.SUPPRESS)
-        super().__init__(*args, **kwargs)
-
-    def __call__(self, parser_, namespace, values, option_string=None):
-        ret = ['__codes__ = {']
-        cv2.imshow('calibration', 0)
-
-        if values == 'all':
-            print('enter backspace')
-            backspace_code = cv2.waitKeyEx()
-            if backspace_code != Codes.backspace:
-                ret.append(f'\t"backspace": {backspace_code},')
-            print('enter esc')
-            esc_code = cv2.waitKeyEx()
-            if esc_code != Codes.esc:
-                ret.append(f'\t"esc": {esc_code}')
-        else:
-            backspace_code = Codes.backspace
-            esc_code = Codes.esc
-
-        for name, v in Codes.__dict__.items():
-            if name.startswith('_'):
-                continue
-            if (not values == 'all') and v < 128:
-                continue
-
-            print(f'enter code for {name}, backspace to not register this key, or esc to quit and save calibration')
-            code = cv2.waitKeyEx()
-            if code == backspace_code:
-                continue
-            elif code == esc_code:
-                break
-            if code == v:
-                print(f'"{name}": {code!r}, same as default')
-            else:
-                ret.append(f'\t"{name}": {code!r},')
-                print(f'"{name}": {code!r}')
-        ret.append('}')
-        with open('__calibration__.py', 'w') as w:
-            w.write('\n'.join(ret))
-        print('calibration saved to __calibration__.py')
-        exit()
-
-
-class DeprecatedAction(argparse.Action):
-    """
-    Informs the user the option has been deprecated
-    """
-
-    def __init__(self, *args, use_instead: str, **kwargs):
-        kwargs.setdefault('help', argparse.SUPPRESS)
-        super().__init__(*args, **kwargs)
-        self.use_instead = use_instead
-
-    def __call__(self, parser_, namespace, values, option_string=None):
-        warnings.warn(DeprecationWarning(
-            f'{option_string} is deprecated, use {self.use_instead} instead, argument ignored'))
-
-
 parser = argparse.ArgumentParser('strider', fromfile_prefix_chars='@',
                                  formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-parser.add_argument('src_path', action='store', help='path to the source video')
-parser.add_argument('tracks_path', action='store', help='path to the trackpack file')
-parser.add_argument('-t', '--tags', action='store', nargs='+', type=str, help='path to a quick tags file',
-                    required=False, default=[], dest='quick_tags_path')
-parser.add_argument('-qt', action='store', nargs='+', type=str, help='quick tags to add', required=False,
-                    default=[], dest='quick_tags')
+parser.add_argument('trackpack_path', action='store', default='?', nargs='?',
+                    help='path to the trackpack file, default is to prompt')
+parser.add_argument('video_path', action='store', default='?', nargs='?',
+                    help='path to the source video, default is to load from trackpack file, or prompt')
 parser.add_argument('--frame_step', action='store', type=int, help='set the regular play speed in frames',
                     default=1, required=False, dest='step')
 parser.add_argument('--seek_step', action='store', type=float,
@@ -103,66 +42,30 @@ parser.add_argument('--line_width', action='store', type=int, help='the width of
 parser.add_argument('--auto_play_wait', action='store', type=int,
                     help='the time to wait between auto-play frames',
                     default=5, required=False, dest='auto_play_wait')
-parser.add_argument('--force_flush', action='store_true', dest='force_flush', required=False,
-                    default=False, help='set to force flushing on every frame (useful for 4k videos)')
 
 # raise is always true in dev mode
 parser.add_argument('--raise', action='store_true', default=strider.__dev__, required=False, dest='raise_',
                     help='raise and quit on exceptions that would normally be caught, use when debugging')
 
 # all options below here exit the program if used
-parser.add_argument('--calibrate', action=Calibrate, nargs='?', choices=['all'],
+parser.add_argument('--calibrate', action=strider.CalibrateAction, nargs='?', choices=['all'],
                     help='run the calibration process and exit,'
                          ' entering --calibrate all will also calibrate the common ascii keys')
 parser.add_argument('--version', action='version', version=strider.__version__)
 
-# deprecated (do not end the program)
-parser.add_argument('--step', action=DeprecatedAction, use_instead='--frame_step')
-parser.add_argument('--back_step', '--backstep', action=DeprecatedAction, use_instead='--seek_step')
-parser.add_argument('--linewidth', action=DeprecatedAction, use_instead='--line_width')
-parser.add_argument('--pointradius', action=DeprecatedAction, use_instead='--point_radius')
 
-
-class Codes:
-    # a static class containing all the key codes the program needs. These values can be overridden
-    # by the calibration file
-    right = 2555904
-    left = 2424832
-    home = 2359296
-    end = 2293760
-
-    # everything below here is ascii
-    esc = 27
-    backspace = 8
-    space = ord(' ')
-    enter = ord('\r')
-    tab = ord('\t')
-
-    a = ord('a')
-    c = ord('c')
-    d = ord('d')
-    g = ord('g')
-    h = ord('h')
-    i = ord('i')
-    n = ord('n')
-    p = ord('p')
-    q = ord('q')
-    r = ord('r')
-    s = ord('s')
-    t = ord('t')
-    u = ord('u')
-    w = ord('w')
-    z = ord('z')
-
-    shift_q = ord('Q')
-    shift_z = ord('Z')
-
-
-def maybe_import(module_name, var_name, var_type=object, default=None, report=True):
+def maybe_import(module_name, var_name, var_type=object, default=None, report=True, **kwargs):
     try:
-        mod = importlib.import_module(module_name)
+        spec = importlib.util.find_spec(module_name)
+        if not spec:
+            return default
     except ImportError:
         return default
+
+    mod = importlib.util.module_from_spec(spec)
+    for k, v in kwargs.items():
+        setattr(mod, k, v)
+    spec.loader.exec_module(mod)
 
     try:
         v = getattr(mod, var_name)
@@ -180,90 +83,43 @@ def maybe_import(module_name, var_name, var_type=object, default=None, report=Tr
     return v
 
 
+tk = None
+tk_filedialog = None
+tk_messagebox = None
+
+
+def init_tk():
+    global tk, tk_filedialog, tk_messagebox
+    if not tk:
+        try:
+            import tkinter as tk
+            import tkinter.filedialog as tk_filedialog
+            import tkinter.messagebox as tk_messagebox
+        except ImportError as e:
+            raise ImportError('cannot import tkinter, source path must be set in arguments') from e
+
+        root = tk.Tk()
+        root.withdraw()
+
+
 def main(args=None):
-    args = parser.parse_args(args)
+    if args is None or isinstance(args, list):
+        args = parser.parse_args(args)
 
     # this global variable is required because the trackbar causing a redraw even when the trackbar moved because of
     # the redraw
     # (since programmatically moving the trackbar also triggers its callback, in true cv2 fashion)
-    suppress_trackbar_seek = False
+    suppress_trackbar_seek = BoolBox()
 
     # to disable clicks while we don't want them (e.g. when entering special commands)
-    suppress_click = False
+    suppress_click = BoolBox()
 
     # to tell if we should even wait for keystroke
     auto_play = False
 
-    user_codes = maybe_import('__calibration__', '__codes__', dict, {})
-    for name, v in user_codes.items():
-        if not hasattr(Codes, name):
-            raise AttributeError(name)
-        setattr(Codes, name, v)
+    codes = maybe_import('__calibration__', '__codes__', strider.Codes, Codes=strider.Codes) or strider.Codes()
 
-    strider.KeyCommand.code_dict = Codes.__dict__
-
-    quick_tags = set(x.lower() for x in args.quick_tags)
-    if args.quick_tags_path:
-        for qtp in args.quick_tags_path:
-            with open(qtp) as r:
-                r = (x.lower().strip() for x in r)
-                quick_tags.update(r)
-    if quick_tags:
-        quick_tags, num = strider.resolve_quick_tags(sorted(quick_tags))
-        print(f'loaded {num} quick tags')
-        if not num:
-            quick_tags = None
-    else:
-        quick_tags = None
-
-    def get_cv_input(prompt):
-        nonlocal suppress_click
-        suppress_click = True
-        print(prompt)
-        edit = strider.LineEdit(report=True)
-        while True:
-            code = cv2.waitKeyEx()
-            if code == Codes.backspace:
-                edit.backspace()
-            elif code == Codes.enter:
-                ret = edit.enter()
-                break
-            elif code == Codes.esc:
-                print('..cancelled')
-                ret = None
-                break
-            elif code == Codes.right:
-                edit.right()
-            elif code == Codes.left:
-                edit.left()
-            elif code == Codes.home:
-                edit.home()
-            elif code == Codes.end:
-                edit.end()
-            elif code < 128:
-                c = chr(code) if code < 0x110000 else ''
-                edit.feed(c)
-        suppress_click = False
-        return ret
-
-    no_default = object()  # singleton to fail on no tags
-
-    def get_quick_tag(additional_codes={}, default=no_default):
-        if not quick_tags:
-            if default is no_default:
-                raise Exception('no quick tags are configured')
-            return default
-        print('QUICK TAGS:')
-        tag_dict = ChainMap(quick_tags, additional_codes)
-        while isinstance(tag_dict, Mapping):
-            for k, m, _ in tag_dict.values():
-                print(f'\t{k}: {m}')
-            code = cv2.waitKeyEx()
-            if code in tag_dict:
-                _, _, tag_dict = tag_dict[code]
-            else:
-                print('code not recognized')
-        return tag_dict
+    strider.KeyCommand.import_codes(codes)
 
     def on_mouse(event, x, y, flags, param):
         if suppress_click:
@@ -297,10 +153,8 @@ def main(args=None):
         next_frame()
 
     def show_frame(frame):
-        nonlocal suppress_trackbar_seek
-        suppress_trackbar_seek = True
-        cv2.setTrackbarPos('position', 'strider', int(view.next_frame_index // view.fps))
-        suppress_trackbar_seek = False
+        with suppress_trackbar_seek:
+            cv2.setTrackbarPos('position', 'strider', int(view.next_frame_index // view.fps))
         cv2.imshow('strider', frame)
 
     def next_frame():
@@ -317,8 +171,8 @@ def main(args=None):
 
     def info_msg():
         d = (
-            ('source video', args.src_path),
-            ('source tracks', args.tracks_path),
+            ('source video', video_path),
+            ('source tracks', pack_path),
             ('next frame', view.next_frame_index),
             ('total frames', view.total_frames),
             ('current time (approx)', ts_to_str(*view.curr_time_approx())),
@@ -328,7 +182,7 @@ def main(args=None):
             ('video fps', view.fps),
             ('view rectangle', repr(view.view_window)),
             ('zoom', 'x' + str(view.real_view.size_ratio(view.view_window))),
-            ('quick tags', ', '.join(strider.tag_names(quick_tags))),
+            ('quick tags', str(quick_tags)),
         )
         return 'INFO:\n' + '\n'.join(f'\t{n}: {v}' for n, v in d)
 
@@ -344,7 +198,7 @@ def main(args=None):
         ))
 
     def report_jump(frame):
-        print(f'jumped to frame {frame}, approx {ts_to_str(*view.approx_frame_to_time(frame,True))}')
+        print(f'jumped to frame {frame}, approx {ts_to_str(*view.approx_frame_to_time(frame, True))}')
 
     @strider.SpecialCommand
     def toggle_track(tid):
@@ -405,6 +259,8 @@ def main(args=None):
             tid = ...
         tag = tag.lower()
         t = view.add_tag(tid, tag=tag)
+        if tag not in quick_tags:
+            quick_tags.add(tag)
         print(f'tag {tag} added to {t}')
 
     @strider.SpecialCommand
@@ -421,6 +277,8 @@ def main(args=None):
             print(f'tag {tag} removed from {t}')
         else:
             print(f'track {t} doesn\'t have tag {tag}')
+        if not track_pack.has_tag(tag):
+            quick_tags.remove(tag)
 
     @strider.SpecialCommand
     def enable_tag(tag):
@@ -436,114 +294,125 @@ def main(args=None):
         view.disable_all(tag)
         print(f'all tracks with tag {tag} disabled (except for active track)')
 
-    @strider.key_command(Codes.space)
+    @strider.key_command(codes.space, doc_placeholders=(args.step,))
     def step_forward():
-        """Step forward one frame, or amount specified in the arguments"""
+        """Step forward {!n:frame}"""
         if next_frame() is None:
             print('end of video')
-        if args.force_flush:  # todo look more into why we need this (for 4k videos)
+        if force_flush:  # todo look more into why we need this (for 4k videos)
             # https://stackoverflow.com/questions/52038222/opencv-python-window-not-refreshing-4k-videos-unless-calling-waitkey1
             cv2.waitKey(1)
 
-    @strider.key_command(Codes.left)
+    @strider.key_command(codes.left, doc_placeholders=(args.seek_step,))
     def step_backwards():
-        """Step backwards a second, or amount specified in the arguments"""
+        """Step backwards {!n:second}"""
         view.back_step()
         next_frame()
 
-    @strider.key_command(Codes.right)
+    @strider.key_command(codes.right, doc_placeholders=(args.seek_step,))
     def step_forwards():
-        """Step forwards a second, or amount specified in the arguments"""
+        """Step forwards {!n:second}"""
         view.fore_step()
         next_frame()
 
-    @strider.key_command(Codes.z, allow_on_auto_play=True)
+    @strider.key_command(codes.z, allow_on_auto_play=True, doc_placeholders=(args.zoom_step,))
     def zoom_in():
-        """Zoom in x2, or amount specified in the arguments"""
-        view.zoom_in(zoom_step)
+        """Zoom in x{}"""
+        view.zoom_in(args.zoom_step)
         this_frame()
 
-    @strider.key_command(Codes.shift_z, allow_on_auto_play=True)
+    @strider.key_command(codes.shift_z, allow_on_auto_play=True, doc_placeholders=(args.zoom_step,))
     def zoom_out():
-        """Zoom out x2, or amount specified in the arguments"""
-        view.zoom_out(zoom_step)
+        """Zoom out x{}"""
+        view.zoom_out(args.zoom_step)
         this_frame()
 
-    @strider.key_command(Codes.esc, allow_on_auto_play=True)
+    @strider.key_command(codes.esc, allow_on_auto_play=True)
     def quit():
         """Exit the program"""
         return True
 
-    @strider.key_command(Codes.a, allow_on_auto_play=True)
+    @strider.key_command(codes.a, allow_on_auto_play=True, doc_placeholders=(args.move_step,))
     def move_left():
-        """Move the view left 10 pixels, or amount specified in the arguments"""
+        """Move the view left {!n:pixel}"""
         view.move_view(x_off=-args.move_step)
         this_frame()
 
-    @strider.key_command(Codes.d, allow_on_auto_play=True)
+    @strider.key_command(codes.d, allow_on_auto_play=True, doc_placeholders=(args.move_step,))
     def move_right():
-        """Move the view right 10 pixels, or amount specified in the arguments"""
+        """Move the view right {!n:pixel}"""
         view.move_view(x_off=args.move_step)
         this_frame()
 
-    @strider.key_command(Codes.w, allow_on_auto_play=True)
+    @strider.key_command(codes.w, allow_on_auto_play=True, doc_placeholders=(args.move_step,))
     def move_up():
-        """Move the view up 10 pixels, or amount specified in the arguments"""
+        """Move the view up {!n:pixel}"""
         view.move_view(y_off=-args.move_step)
         this_frame()
 
-    @strider.key_command(Codes.s, allow_on_auto_play=True)
+    @strider.key_command(codes.s, allow_on_auto_play=True, doc_placeholders=(args.move_step,))
     def move_down():
-        """Move the view down 10 pixels, or amount specified in the arguments"""
+        """Move the view down {!n:pixel}"""
         view.move_view(y_off=args.move_step)
         this_frame()
 
-    @strider.key_command(Codes.n)
+    @strider.key_command(codes.n)
     def create_new_track():
         """Create and activate a new track with a semi-random id"""
         new_track(...)
 
-    @strider.key_command(Codes.g)
-    def assign_quick_tag():
+    @strider.key_command(codes.g)
+    def assign_tag_input():
         """Assign a tag to the active tag, chosen from the quick tags"""
-        if not quick_tags:
-            print('no quick tags, use -t ro configure quick tags')
-        elif not view.active_track:
+        if not view.active_track:
             print('no active track')
         else:
-            t = get_quick_tag()
+            with suppress_click:
+                t = quick_tags.input('type tag (<tab>-autocomplete, <esc>-cancel)', codes, report=True)
             if t:
                 tag(t)
-            else:
-                print('cancelled')
 
-    @strider.key_command(Codes.h)
+    @strider.key_command(codes.shift_g)
+    def remove_tag_input():
+        """Assign a tag to the active tag, chosen from the quick tags"""
+        if not view.active_track:
+            print('no active track')
+        else:
+            with suppress_click:
+                t = quick_tags.input('type tag (<tab>-autocomplete, <esc>-cancel)', codes, report=True)
+            if t:
+                remove_tag(t)
+
+    @strider.key_command(codes.h)
     def help():
         """Display this help message"""
         print('HELP:\n', textwrap.indent(help_msg(), prefix='\t'))
 
-    @strider.key_command(Codes.i)
+    @strider.key_command(codes.i)
     def info():
         """Show general variable information about the strider environment"""
         print(info_msg())
 
-    @strider.key_command(Codes.t)
+    @strider.key_command(codes.t)
     def list_tracks():
         """List all the tracks, and some information about them"""
         ret = [f'TRACKS ({len(view.track_pack.tracks)} total):']
         for t in view.tracks_sorted():
+            if isinstance(t, str):
+                ret.append(t + ':')
+                continue
             stats = t.stats(enabled=str(view.track_pack.is_enabled(t)), active=view.active_track is t)
             ret.append(f'\t{stats}')
         print('\n'.join(ret))
 
-    @strider.key_command(Codes.r)
+    @strider.key_command(codes.r)
     def jump_to_zero():
         """Jump to the first frame of the video"""
         view.reset()
         report_jump(0)
         next_frame()
 
-    @strider.key_command(Codes.home)
+    @strider.key_command(codes.home)
     def jump_to_start():
         """Jump to the start of the current track"""
         if not view.active_track:
@@ -557,7 +426,7 @@ def main(args=None):
                 report_jump(span[0])
                 next_frame()
 
-    @strider.key_command(Codes.end)
+    @strider.key_command(codes.end)
     def jump_to_end():
         """Jump to the end of the active track"""
         if not view.active_track:
@@ -571,14 +440,14 @@ def main(args=None):
                 report_jump(span[1])
                 next_frame()
 
-    @strider.key_command(Codes.p)
+    @strider.key_command(codes.p)
     def save_tracks():
         """Save the tracks to the designated json file"""
-        with open(args.tracks_path, 'w') as w:
+        with open(pack_path, 'w') as w:
             view.track_pack.write(w)
         print('saved!')
 
-    @strider.key_command(Codes.u)
+    @strider.key_command(codes.u)
     def undo():
         """Remove last point (up to current frame) in the active track"""
         if not view.active_track:
@@ -588,11 +457,12 @@ def main(args=None):
             this_frame()
             print(f'point deleted: {deleted}')
 
-    @strider.key_command(Codes.q)
+    @strider.key_command(codes.q)
     def batch_enable():
         """Enable tracks by a specified quick tag (or space to enable all)"""
-        t = get_quick_tag({Codes.space: ('space', '<all>', ...)}, default=...)
-        if t is ...:
+        with suppress_click:
+            t = quick_tags.input('type tag (<tab>-autocomplete, <esc>-cancel, blank for all)', codes, report=True)
+        if t is '':
             view.enable_all()
             print('all tracks enabled')
         elif t is None:
@@ -601,11 +471,12 @@ def main(args=None):
             enable_tag(t)
         this_frame()
 
-    @strider.key_command(Codes.shift_q)
+    @strider.key_command(codes.shift_q)
     def batch_disable():
         """Disable tracks by quick tag (or space to enable all), except the active track, if it exists"""
-        t = get_quick_tag({Codes.space: ('space', '<all>', ...)}, default=...)
-        if t is ...:
+        with suppress_click:
+            t = quick_tags.input('type tag (<tab>-autocomplete, <esc>-cancel, blank for all)', codes, report=True)
+        if t is '':
             view.disable_all()
             print('all tracks disabled (except for active track)')
         elif t is None:
@@ -614,7 +485,7 @@ def main(args=None):
             disable_tag(t)
         this_frame()
 
-    @strider.key_command(Codes.tab)
+    @strider.key_command(codes.tab)
     def start_auto_play():
         """Enable auto-play, will continue the video until a key is pressed.
          Some keys (zoom, move zoomed rectangle, ...) will not stop the auto-play."""
@@ -622,10 +493,13 @@ def main(args=None):
         auto_play = True
         print("started auto-play, press tab again to end")
 
-    @strider.key_command(Codes.c)
+    @strider.key_command(codes.c)
     def special_command():
         """Run a special command as typed:"""
-        line = get_cv_input('enter command:')
+        with suppress_click:
+            line_edit = strider.LineEdit(report=True, autocomplete=strider.SpecialCommand.keys())
+            line = line_edit.cv_input('enter command (tab to autocomplete, esc to cancel)', codes)
+
         if line:
             try:
                 comm, arguments = strider.SpecialCommand.parse_func_call(line)
@@ -641,12 +515,59 @@ def main(args=None):
 
     special_command.__doc__ += ''.join('\n\t' + sc.__doc__ for sc in strider.SpecialCommand.values())
 
-    zoom_step = args.zoom_step
+    pack_path = args.trackpack_path
+    if pack_path == '?':
+        init_tk()
+        pack_path = tk_filedialog.asksaveasfilename(
+            defaultextension='json', filetypes=['json {.json .stp}', '{all files} *'],
+            title='choose a trackpack file to use', confirmoverwrite=False
+        )
+        if not pack_path:
+            print('cancelled')
+            exit()
 
-    view = strider.StriderView(pack_path=args.tracks_path, video_source_path=args.src_path,
+    try:
+        with open(pack_path) as r:
+            track_pack = TrackPack.read(r)
+    except (json.JSONDecodeError, KeyError) as e:
+        raise Exception('could not open file') from e
+    except FileNotFoundError:
+        print(f'new file created: {pack_path}')
+        track_pack = TrackPack(name=pack_path)
+
+    quick_tags = strider.QuickTagRepo(track_pack.all_tags())
+
+    video_path = args.video_path
+    if video_path == '?':
+        if track_pack.video_path and path.exists(track_pack.video_path):
+            video_path = track_pack.video_path
+        else:
+            init_tk()
+            video_path = tk_filedialog.askopenfilename(
+                filetypes=['video {.mp4 .mov .avi .mkv}', '{all files} *'], title='choose a video file to open'
+            )
+            if not video_path:
+                print('cancelled')
+                exit()
+
+    if track_pack.video_path != video_path:
+        try:
+            init_tk()
+        except ImportError:
+            pass
+        else:
+            if tk_messagebox.askyesno(
+                    'set default source',
+                    "the selected video is different from the trackpack's default,"
+                    " would you like to set it as default?"):
+                track_pack.video_path = video_path
+
+    view = strider.StriderView(track_pack=track_pack, video_source_path=video_path,
                                play_step_frame=args.step, seek_step_sec=args.seek_step,
                                line_width=args.line_width, point_radius=args.point_radius)
     view.track_pack.enable_all()
+
+    force_flush = view.real_view.breaks_bounds(Rectangle(0, 0, 2000, 1100))
 
     cv2.namedWindow('strider', cv2.WINDOW_NORMAL)
     cv2.setMouseCallback('strider', on_mouse)